Moved test to a separate mod. Added a fail signature scenario.
+ Cargo fmt / check / clippy
