+++ massa-signature/Cargo.toml
@@ -7,11 +7,9 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-anyhow = "1.0"
 schnorrkel = "0.10.2"
 merlin = "3.0.0"
 transition = { git = "https://github.com/AurelienFT/transition" }
-tokio = { version = "1.23", features = ["full"] }
 
 bs58 = { version = "0.4", features = ["check"] }
 displaydoc = "0.2"
@@ -27,3 +25,5 @@ massa_serialization = { path = "../massa-serialization" }
 [dev-dependencies]
 serial_test = "0.10"
 serde_json = "1.0"
+anyhow = "1.0"
+tokio = { version = "1.23", features = ["full"] }
